+++ src/main/java/com/vita/sjk/zhihudaily/constants/Constants.java
@@ -6,5 +6,5 @@ package com.vita.sjk.zhihudaily.constants;
 public class Constants {
 
     public static final String NEWS_ID = "id";
-    public static final int NEWS_ID_INVALID = -1;
+    public static final long NEWS_ID_INVALID = -1;
 }
